Settings UI: move Sitemaps into its own card
