+++ _inc/client/traffic/verification-services.jsx
@@ -2,11 +2,9 @@
  * External dependencies
  */
 import React from 'react';
-import { connect } from 'react-redux';
 import { translate as __ } from 'i18n-calypso';
 import TextInput from 'components/text-input';
 import ExternalLink from 'components/external-link';
-import get from 'lodash/get';
 
 /**
  * Internal dependencies
@@ -16,19 +14,13 @@ import {
 	FormLabel
 } from 'components/forms';
 import { ModuleSettingsForm as moduleSettingsForm } from 'components/module-settings/module-settings-form';
-import { ModuleToggle } from 'components/module-toggle';
 import SettingsCard from 'components/settings-card';
 import SettingsGroup from 'components/settings-group';
-import { getSiteAdminUrl, isSiteVisibleToSearchEngines } from 'state/initial-state';
 
 export const VerificationServices = moduleSettingsForm(
 	React.createClass( {
-
 		render() {
-			let verification     = this.props.getModule( 'verification-tools' ),
-				sitemaps         = this.props.getModule( 'sitemaps' ),
-				sitemap_url      = get( sitemaps, [ 'extra', 'sitemap_url' ], '' ),
-				news_sitemap_url = get( sitemaps, [ 'extra', 'news_sitemap_url' ], '' );
+			const verification = this.props.getModule( 'verification-tools' );
 			return (
 				<SettingsCard
 					{ ...this.props }
@@ -115,51 +107,9 @@ export const VerificationServices = moduleSettingsForm(
 							}
 						</FormFieldset>
 					</SettingsGroup>
-					<SettingsGroup support={ sitemaps.learn_more_button }>
-						<ModuleToggle
-							slug="sitemaps"
-							compact
-							activated={ this.props.getOptionValue( 'sitemaps' ) }
-							toggling={ this.props.isSavingAnyOption( 'sitemaps' ) }
-							toggleModule={ this.props.toggleModuleNow }>
-							{ __( 'Generate XML sitemaps' ) }
-						</ModuleToggle>
-						{
-							this.props.isSiteVisibleToSearchEngines
-								? this.props.getOptionValue( 'sitemaps' ) && (
-									<FormFieldset>
-										<p>
-											<ExternalLink icon={ true } target="_blank" href={ sitemap_url }>{ sitemap_url }</ExternalLink>
-											<br />
-											<ExternalLink icon={ true } target="_blank" href={ news_sitemap_url }>{ news_sitemap_url }</ExternalLink>
-										</p>
-										<p className="jp-form-setting-explanation">{ __( 'Your sitemap is automatically sent to all major search engines for indexing.' ) }</p>
-									</FormFieldset>
-								)
-								: (
-									<p className="jp-form-setting-explanation">
-										{
-											__( 'Your site is not currently accessible to search engines. You might have "Search Engine Visibility" disabled in your {{a}}Reading Settings{{/a}}.', {
-												components: {
-													a: <a href={ this.props.siteAdminUrl + 'options-reading.php' } />
-												}
-											} )
-										}
-									</p>
-								)
-						}
-					</SettingsGroup>
 				</SettingsCard>
 			);
 		}
 	} )
 );
 
-export default connect(
-	state => {
-		return {
-			isSiteVisibleToSearchEngines: isSiteVisibleToSearchEngines( state ),
-			siteAdminUrl: getSiteAdminUrl( state )
-		};
-	}
-)( VerificationServices );
